backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
PyJWT==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
# PostgreSQL adapter – install based on Python version:
#   Python 3.13+:  pip install "psycopg[binary]"
#   Python <=3.12: pip install psycopg2-binary==2.9.10
